# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nom de votre API",
        "description": "Description de votre API",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Liste des produits",
                "description": "Renvoie la liste de tous les produits avec leurs détails et catégories.",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "Liste des produits récupérée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID du produit.",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Nom du produit.",
                                                "type": "string"
                                            },
                                            "price": {
                                                "description": "Prix du produit.",
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "stock": {
                                                "description": "Stock disponible du produit.",
                                                "type": "integer"
                                            },
                                            "image": {
                                                "description": "URL de l'image du produit.",
                                                "type": "string"
                                            },
                                            "categories": {
                                                "description": "Catégories auxquelles appartient le produit.",
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Produits"
                ],
                "summary": "Créer un nouveau produit",
                "description": "Crée un nouveau produit avec les détails fournis.",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom du produit.",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Prix du produit.",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "stock": {
                                        "description": "Stock disponible du produit.",
                                        "type": "integer"
                                    },
                                    "categories": {
                                        "description": "ID des catégories auxquelles appartient le produit.",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "imageinput": {
                                        "description": "Image du produit (format: pdf, jpg, png, max: 2MB).",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit créé avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Afficher les détails d'un produit",
                "description": "Récupère les détails d'un produit spécifié par son ID.",
                "operationId": "b46f4aa73232d0368eee698a082342a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit à afficher.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du produit récupérés avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID du produit.",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Nom du produit.",
                                            "type": "string"
                                        },
                                        "price": {
                                            "description": "Prix du produit.",
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "stock": {
                                            "description": "Stock disponible du produit.",
                                            "type": "integer"
                                        },
                                        "image": {
                                            "description": "URL de l'image du produit.",
                                            "type": "string"
                                        },
                                        "categories": {
                                            "description": "Catégories auxquelles appartient le produit.",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produit non trouvé."
                    }
                }
            },
            "put": {
                "tags": [
                    "Produits"
                ],
                "summary": "Mettre à jour un produit",
                "description": "Met à jour les détails d'un produit spécifié par son ID.",
                "operationId": "1419d479f9340e30e67fe21a2ca525b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit à mettre à jour.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom du produit.",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Prix du produit.",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "stock": {
                                        "description": "Stock disponible du produit.",
                                        "type": "integer"
                                    },
                                    "categories": {
                                        "description": "ID des catégories auxquelles appartient le produit.",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "imageinput": {
                                        "description": "Image du produit (format: pdf, jpg, png, max: 2MB).",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit mis à jour avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produit non trouvé."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produits"
                ],
                "summary": "Supprimer un produit",
                "description": "Supprime un produit spécifié par son ID.",
                "operationId": "fe34b364a3f192305598b65d04d06b5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit à supprimer.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produit supprimé avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Votre produit a bien été supprimé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produit non trouvé."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "description": "ID du produit",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nom du produit",
                        "type": "string"
                    },
                    "price": {
                        "description": "Prix du produit",
                        "type": "number",
                        "format": "float"
                    },
                    "stock": {
                        "description": "Stock du produit",
                        "type": "integer"
                    },
                    "image": {
                        "description": "URL de l'image du produit",
                        "type": "string",
                        "format": "url"
                    },
                    "created_at": {
                        "description": "Date de création du produit",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour du produit",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}